backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.5.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.1
coverage==7.10.7
cryptography==44.0.2
deprecation==2.1.0
Flask==3.1.0
frozenlist==1.5.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.3.0
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.2
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.4.0
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
